feat: refactor function bindgen
